third_party/WebKit/LayoutTests/external/wpt/infrastructure/testdriver/actions/elementPosition-expected.txt

This is a testharness.js-based test.
FAIL TestDriver actions: element position assert_array_equals: property 0, expected 50 but got 0
Harness: the test ran to completion.
